requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
Flask==3.0.3
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pycparser==2.22
pyodbc==5.1.0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.0
rpds-py==0.18.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
